Fix addition of zero values during pushes and logic error.
This was otherwise a bug where reading large values, with constituent
bits spanning multiple bytes, could lead to a situation where partial
reads resulted in a zero value for that read alone, which otherwise
would properly shift the accumulated value to generate the correct
result, but were getting silently dumped on the floor.

As well, the push method for LittleEndian was improperly shifting the
value, instead of the accumulator, with the accumulated bits instead of
the bits for the given value.