From{Bit,Byte}StreamWith: allow lifetime-constrained Contexts
In current implementation, `From{Bit,Byte}StreamWith` impose that the `Context`
be static. In some cases, we want to pass a `Context` which owns a reference to
another struct.

This commit adds a lifetime to these traits to allow using a `Context`
constrained by a lifetime. Ex.:

```rust
impl FromBitStreamWith<'a> for FrameHeader {
    type Context = StreamInfo<'a>;
    [...]
}
```

Implementations for which the `Context` is `'static` bound can use:

```rust
impl FromBitStreamWith<'_> for FrameHeader {
    type Context = StreamInfo;
    [...]
}
```