Pull all the unsigned numerics into their own trait
That is, while both UnsignedNumeric and SignedNumeric
are supertraits of Numeric,
BitRead::read() and friends only returns UnsignedNumeric
(and BitWrite::write() only writes UnsignedNumeric)
which prevents things like:

bitwriter.write(8, -1i8)

and forces the use of:

bitwriter.write_signed(8, -1i8)

at compile-time at the type level.

Because the signed readers/writers are implemented in terms
of their unsigned variants, the signed/unsigned traits have
conversion methods to switch between each other.

Also because this could be a significant source of breakage,
I'll increment the major version number and take a bit of time
before pushing this out to Crates.io