Add an Integer trait to encompass signed and unsigned integer types
That is, read(), read_to(), write() and write_out() take Integer,
which is implemented by signed and unsigned ints which delegate
their calls to (read|write)_(unsigned|signed) depending on their
signedness.

This makes the "unsigned" calls the base implementations
the rest of the traits are based off of.

It should also minimize the amount of code breakage going forward.