Fix some performance regressions
Update method naming to ensure that pop_bits is for
values in the queue aligned to the correct endianness,
while push_value is for building final values.

I can't just reverse these to build a bit writer,
because popping whole bytes off a writer requires
the type to be written to be aligned properly beforehand.