Major overhaul to BitRead trait
Essentially, the old *_in methods ("read_in", "read_unsigned_in", etc.)
have been promoted to the main methods ("read", "read_unsigned", etc.)
while the methods that take a variable number of bits
are suffixed with "_var".

The methods that take a constant number of bits should be
preferred, since file formats are typically implemented
with fields that take a fixed number of bits most of the time.
These methods also validate their bit counts at compile-time,
so they should be slightly more performant.

But because this is a big breaking change,
I've added a "BitRead2" trait that's compatible with the
older calling convention.  A blanket implementation of
"BitRead2" is provided for anything that implements "BitRead"
which simply forwards along calls to the proper method.