Widen out BitCount to (technically) support u32::MAX bits on read/writes
This is absurd in practice, because no primitive type gets anywhere
near that large.

In practice, BitCounts as a concept are for things like FLAC
Rice codes where the number of least-significant bits in each
code are stored as a 4 (or 5) bit value read beforehand.
Because we know the amount of bits we're reading in each
code doesn't get above 15 (or 31), we can forego checking
whether the amount of bits is larger than a u32 can hold.