Optimize short bit reads just a bit further
This smooshes the process of pulling in a new byte,
combining it with the existing partial byte for a final value,
and re-updating the queue with the remainder, into single
hyper-dense expression that runs just a little faster than before.

The multi-byte case could be optimized further, but it's
best to break just one thing at a time.