Add new SignedBitCount type
This separates the process of extracting the sign bit
from a bit count from the process of reading or writing
signed values with that count.

It also allows the signed reading and writing methods
to all funnel into one single method per endinanness
instead of having multiple methods doing almost the same thing.