Optimize bytes reading/writing for primitives
That is, when reading/writing an un-aligned whole primitive value
(like a u32, i64, etc.), reserve a smaller amount of space
for the chunk buffer.

It doesn't make much sense to clear out a whole 1024 bytes
of space for primitives that are mostly 16 bytes or fewer.
In practice, these kinds of primitives tend to be byte-aligned
the vast majority of the time, but I'd like to avoid any
unpleasant performance surprises down the line.