Preliminary addition of a CheckedUnsigned type framework
Normally, when writing values to a stream, we not only need
to verify that number of bits written isn't larger than the type,
but also that the value isn't larger than the number of bits can hold.
For instance, writing 9 bits from a u8 is an error, but also
trying to write a value of 255 in 7 bits.

Using BitCounts enables us to avoid the first check
by guaranteeing the maximum bits isn't larger than our type at
compile-time.

Using CheckedUnsigned enables us to avoid the second check
by guaranteeing the value isn't larger than the number of
bits to be written - either by carrying that information
from a read call or by performing that check when
building a CheckedUnsigned value.

It also eliminates a bit of duplicate bounds checking
by consolidating it into one type.