Divide the Checkable trait into a read-specific one
The CheckableFromBitstream trait is intended for primitives,
but it's possible to envision expanding the Checkable
trait for collections and other things which might not
fit into the regular reading paradigm.  So I'll give myself
the option to expand it later.

The names are still a lot, though, so I'm not entirely sure
whether to try to find something shorter and more descriptive.